Migrated CDD KVS to CDD 0.8.50
