+++ functions/CDDKVSJava/build.gradle
@@ -41,7 +41,7 @@ repositories {
     maven { url 'https://jitpack.io' }
 }
 
-def cddVersion = '0.8.49'
+def cddVersion = '0.8.50'
 def slf4jVersion = '1.7.30'
 def jacksonVersion = '2.10.2'
 def awsLambdaJavaCoreVersion = '1.2.0'
