+++ functions/CDDKVSJava/function.conf
@@ -5,22 +5,22 @@ conf {
   functionName = "CDDKVSJava"
   // This must be the full package name of the App class.  The JSON and binary handlers are
   //   implemented in BaselineAppInterface, you do not need to implement them.
-  // For JSON use   - handlerName = "com.amazonaws.greengrass.cddkinesis.App::"${conf.javaJsonHandler}
-  // For binary use - handlerName = "com.amazonaws.greengrass.cddkinesis.App::"${conf.javaBinaryHandler}
-  handlerName = "com.amazonaws.greengrass.cddkinesis.App::"${conf.javaJsonHandler}
+  // For JSON use   - handlerName = "com.amazonaws.greengrass.cddkvs.App::"${conf.javaJsonHandler}
+  // For binary use - handlerName = "com.amazonaws.greengrass.cddkvs.App::"${conf.javaBinaryHandler}
+  handlerName = "com.amazonaws.greengrass.cddkvs.App::"${conf.javaJsonHandler}
   // This can be any alias name you'd like
   aliasName = "PROD"
   memorySizeInKb = 131072
   pinned = true
   timeoutInSeconds = 60
   // Messages published on this topic to AWS IoT in the same region will be sent to this function
-  fromCloudSubscriptions = [${AWS_IOT_THING_NAME}"/cdd/kinesis/input"]
+  fromCloudSubscriptions = [${AWS_IOT_THING_NAME}"/cdd/kvs/input"]
   // Messages published by this function on this topic will be sent to AWS IoT in the same region
-  toCloudSubscriptions = [${AWS_IOT_THING_NAME}"/cdd/kinesis/output"]
+  toCloudSubscriptions = [${AWS_IOT_THING_NAME}"/cdd/kvs/output"]
   // Messages published by this function will be routed to other functions on the same core with the these input topics
-  outputTopics = [${AWS_IOT_THING_NAME}"/cdd/kinesis/output"]
+  outputTopics = [${AWS_IOT_THING_NAME}"/cdd/kvs/output"]
   // Messages published by other functions on the same core with these output topics will be routed to this function
-  inputTopics = [${AWS_IOT_THING_NAME}"/cdd/kinesis/input"]
+  inputTopics = [${AWS_IOT_THING_NAME}"/cdd/kvs/input"]
   localDeviceResources = [
       {
         path = "/dev/video0"
