+++ functions/CDDKVSJava/src/main/java/com/amazonaws/greengrass/cddkvs/AppInjector.java
@@ -1,0 +1,17 @@
+package com.amazonaws.greengrass.cddkvs;
+
+import com.amazonaws.greengrass.cddkvs.handlers.InputHandler;
+import com.amazonaws.greengrass.cddkvs.handlers.StartupHandler;
+import com.awslabs.aws.iot.greengrass.cdd.BaselineInjector;
+import com.awslabs.aws.iot.greengrass.cdd.modules.BaselineAppModule;
+import dagger.Component;
+
+import javax.inject.Singleton;
+
+@Singleton
+@Component(modules = {BaselineAppModule.class, AppModule.class})
+public interface AppInjector extends BaselineInjector {
+    StartupHandler startupHandler();
+
+    InputHandler inputHandler();
+}
